requirements_streamlit.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
google-adk>=0.1.0
google-genai>=0.3.0
asyncio
typing-extensions>=4.0.0 